Restructured folder names to adhere to JSX naming standards
